+++ NEWS.md
@@ -1,5 +1,17 @@
-# MsBackendMassIVE 0.1
+# MsBackendMassIVE 0.1.2
 
-- Add MassIVE utilities
-- Add functionality to download and cache MS data files from MassIVE
-- Add unit tests
+- Add `massive_param_file()` to parse params.xml file.
+- Add additional test for `massive_list_files()` and `.massive_data_files()`
+
+# MsBackendMassIVE 0.1.1
+
+- Add package vignette.
+- Replace interactive `readline()` with `overwrite` parameter in
+  `massive_download_file()`.
+- Replace `cat()` with `message()` in `massive_download_file()`.
+
+# MsBackendMassIVE 0.1.0
+
+- Add MassIVE utilities.
+- Add functionality to download and cache MS data files from MassIVE.
+- Add unit tests.